Refactor: Simplify imports by consolidating and standardizing module paths
- Updated imports to use `utils` module consistently across configurations.
- Removed unused `ValidationError` imports for cleaner dependencies.
- Streamlined import paths in `StaticFilesPluginSettings`, `SpaceStatePluginConfig`, and `smib.config`.
